Allow user to specify a maximum channel count
This helps work around the Frsky ppm bug as discussed here:

http://diydrones.com/profiles/blogs/why-frsky-cppm-signal-is-so-disappointing

Any time a channel number > than the maximum user specified channel
count is detected, this channel is flagged as an error and the channel
numbers start over.  In the frsky case, the 8th channel gets truncated,
but it's otherwise a mostly usable signal.

Example:  http://i.imgur.com/na1Ytfc.png

In the wide format CSV export, any frame that was reset because of this
condition is flagged as an error:

    Time [s],ch1,ch2,ch3,ch4,ch5,ch6,ch7,ch8,error
    [...]
    0.918427375000000,1147,1503,1801,1503,1503,1503,1503,1503,
    0.936431500000000,1147,1503,1802,1503,1503,1503,1503,1503,
    0.954428250000000,1147,1503,1802,1503,1503,1503,1503,1503,
    0.972433000000000,1148,1503,1802,1503,1503,1503,1503,1759,
    0.990429125000000,2183,2184,2184,2184,2184,2185,2185,2015,
    1.007737875000000,2183,2184,2184,2184,2185,2185,2185,2015,error
    1.025740625000000,2184,2184,2184,2184,2184,2185,2185,2015,error
    1.043740125000000,2183,2184,2184,2184,2185,2185,2185,2015,error
    1.061741500000000,2183,2184,2184,2184,2185,2185,2185,2015,error
    1.079740875000000,2183,2184,2184,2184,2184,2185,2185,2015,error

In the long format CSV export, the actual value is emitted, but
shown as an error:

    Time [s],val,chan,error
    0.981397500000000,1503,7,
    0.982901125000000,1759,8,
    0.990429125000000,2183,1,
    0.992613750000000,2184,2,
    0.994798250000000,2184,3,
    0.996983000000000,2184,4,
    0.999167750000000,2184,5,
    1.001352750000000,2185,6,
    1.003537750000000,2185,7,
    1.005722750000000,2015,8,
    1.007737875000000,694,1,error
    1.008432000000000,2183,1,
    1.010616625000000,2184,2,
    1.012801000000000,2184,3,